added doubles, fixed small bug in driver

+++ vector.c
@@ -7,7 +7,7 @@
 
 int main (void) {
    vec32 *id1 = initIdentity(5);
-   int32_t **a2 = init2dArr(5,5);
+   double **a2 = init2dArr(5,5);
    a2[1][1] = 5;
    vec32 *vec2 = initVec(a2,5,5);
    printVec(multiply(id1,vec2));
